[REF] core: move tools' tests from base to test_tools
