+++ odoo/addons/base/tests/__init__.py
@@ -3,23 +3,16 @@
 from . import common
 from . import test_acl
 from . import test_api
-from . import test_barcode
 from . import test_base
 from . import test_basecase
 from . import test_cache
-from . import test_intervals
-from . import test_date_utils
 from . import test_deprecation
 from . import test_db_cursor
 from . import test_display_name
 from . import test_expression
-from . import test_float
 from . import test_format_address_mixin
-from . import test_func
 from . import test_groups
 from . import test_http_case
-from . import test_i18n
-from . import test_image
 from . import test_install
 from . import test_avatar_mixin
 from . import test_init
@@ -37,9 +30,7 @@ from . import test_ir_sequence
 from . import test_ir_sequence_date_range
 from . import test_ir_embedded_actions
 from . import test_ir_default
-from . import test_mail
 from . import test_menu
-from . import test_misc
 from . import test_module
 from . import test_module_graph
 from . import test_orm
@@ -52,8 +43,6 @@ from . import test_res_config
 from . import test_res_lang
 from . import test_search
 from . import test_split_table
-from . import test_sql
-from . import test_translate
 # from . import test_uninstall  # loop
 from . import test_user_has_group
 from . import test_views
@@ -70,12 +59,7 @@ from . import test_test_suite
 from . import test_tests_tags
 from . import test_transactions
 from . import test_form_create
-from . import test_cloc
-from . import test_profiler
-from . import test_pdf
 from . import test_neutralize
 from . import test_config_parameter
 from . import test_ir_module_category
-from . import test_configmanager
-from . import test_signature
 from . import test_import_files
