+++ odoo/addons/test_tools/__manifest__.py
@@ -1,0 +1,8 @@
+{
+    'name': 'Test Tools',
+    'category': 'Hidden/Tests',
+    'description': """Tests the Tools.""",
+    'depends': ['base'],
+    'author': 'Odoo S.A.',
+    'license': 'LGPL-3',
+}
