+++ odoo/addons/test_tools/tests/__init__.py
@@ -1,0 +1,17 @@
+from . import test_barcode
+from . import test_cloc
+from . import test_config
+from . import test_date_utils
+from . import test_float_utils
+from . import test_func
+from . import test_i18n
+from . import test_image
+from . import test_intervals
+from . import test_mail
+from . import test_misc
+from . import test_pdf
+from . import test_profiler
+from . import test_safe_eval
+from . import test_set_expression
+from . import test_sql
+from . import test_translate
