+++ odoo/addons/test_tools/tests/config/16.0.conf
