+++ odoo/addons/test_tools/tests/config/cli
