+++ odoo/addons/test_tools/tests/config/empty.conf
