+++ odoo/addons/test_tools/tests/config/save_posix.conf
