+++ odoo/addons/test_tools/tests/test_barcode.py
