+++ odoo/addons/test_tools/tests/test_func.py
