+++ odoo/addons/test_tools/tests/test_sql.py
